+++ main.dart
@@ -1,0 +1,27 @@
+import 'package:logger_console/logger_console.dart';
+
+void main() {
+  Console.log("string or object...");
+
+  Console.log([
+    {"color": "red", "value": "#f00"},
+    {"color": "green", "value": "#0f0"},
+    {"color": "blue", "value": "#00f"},
+    {"color": "cyan", "value": "#0ff"},
+    {"color": "magenta", "value": "#f0f"},
+    {"color": "yellow", "value": "#ff0"},
+    {"color": "black", "value": "#000"},
+  ]);
+
+  Console.group('group data');
+  Console.log("data", [
+    {"color": "red", "value": "#f00"},
+    {"color": "green", "value": "#0f0"},
+    {"color": "blue", "value": "#00f"},
+    {"color": "cyan", "value": "#0ff"},
+    {"color": "magenta", "value": "#f0f"},
+    {"color": "yellow", "value": "#ff0"},
+    {"color": "black", "value": "#000"},
+  ]);
+  Console.groupEnd();
+}